+++ src/test/resources/testNoAgent.Jenkinsfile
@@ -1,0 +1,13 @@
+pipeline {
+    agent none
+
+    stages {
+        stage('Build') {
+            steps {
+                echo "before skip"
+                scmSkip(skipPattern: '.*\\[(ci skip|skip ci)\\].*')
+                echo "after skip"
+            }
+        }
+    }
+}
